Rename files to match snake case notation
