+++ backend/controller/input_validator/input_validator.py
@@ -1,5 +1,5 @@
-from controller.input_validator.ValidationResult import ValidationResult
-from controller.input_validator.ValidationStatus import ValidationStatus
+from controller.input_validator.validation_result import ValidationResult
+from controller.input_validator.validation_status import ValidationStatus
 
 
 class InputValidator:
