+++ backend/service/user_service.py
@@ -1,6 +1,6 @@
-from controller.factory.UserFactory import UserFactory
-from controller.input_validator.UserDataValidator import UserDataValidator
-from controller.input_validator.ValidationStatus import ValidationStatus
+from controller.factory.user_factory import UserFactory
+from controller.input_validator.user_data_validator import UserDataValidator
+from controller.input_validator.validation_status import ValidationStatus
 from model.repository.user_repository import UserRepository
 from model.request_result import RequestResult
 from model.user.role import UserRole
